+++ internal/observability/middleware_test.go
@@ -176,7 +176,7 @@ func TestMiddlewareLogging(t *testing.T) {
 		RoutingDelegate: "routing-delegate",
 		Body:            strings.NewReader("body"),
 	}
-	sreq := &transport.StreamRequest{Meta: req.ToRequestMeta()}
+
 	failed := errors.New("fail")
 
 	baseFields := func() []zapcore.Field {
@@ -268,7 +268,7 @@ func TestMiddlewareLogging(t *testing.T) {
 			},
 		})
 
-		getLog := func() observer.LoggedEntry {
+		getLog := func(t *testing.T) observer.LoggedEntry {
 			entries := logs.TakeAll()
 			require.Equal(t, 1, len(entries), "Unexpected number of logs written.")
 			e := entries[0]
@@ -305,7 +305,7 @@ func TestMiddlewareLogging(t *testing.T) {
 				},
 				Context: logContext,
 			}
-			assert.Equal(t, expected, getLog(), "Unexpected log entry written.")
+			assert.Equal(t, expected, getLog(t), "Unexpected log entry written.")
 		})
 		t.Run(tt.desc+", unary outbound", func(t *testing.T) {
 			res, err := mw.Call(context.Background(), req, newOutbound(tt))
@@ -326,7 +326,7 @@ func TestMiddlewareLogging(t *testing.T) {
 				},
 				Context: logContext,
 			}
-			assert.Equal(t, expected, getLog(), "Unexpected log entry written.")
+			assert.Equal(t, expected, getLog(t), "Unexpected log entry written.")
 		})
 
 		// Application errors aren't applicable to oneway and streaming
@@ -350,7 +350,7 @@ func TestMiddlewareLogging(t *testing.T) {
 				},
 				Context: logContext,
 			}
-			assert.Equal(t, expected, getLog(), "Unexpected log entry written.")
+			assert.Equal(t, expected, getLog(t), "Unexpected log entry written.")
 		})
 		t.Run(tt.desc+", oneway outbound", func(t *testing.T) {
 			ack, err := mw.CallOneway(context.Background(), req, newOutbound(tt))
@@ -371,50 +371,398 @@ func TestMiddlewareLogging(t *testing.T) {
 				},
 				Context: logContext,
 			}
-			assert.Equal(t, expected, getLog(), "Unexpected log entry written.")
-		})
-		t.Run(tt.desc+", stream inbound", func(t *testing.T) {
-			stream, err := transport.NewServerStream(&fakeStream{ctx: context.Background(), request: sreq})
-			require.NoError(t, err)
-			err = mw.HandleStream(stream, newHandler(tt))
-			checkErr(err)
-			logContext := append(
-				baseFields(),
-				zap.String("direction", string(_directionInbound)),
-				zap.String("rpcType", "Streaming"),
-			)
-			logContext = append(logContext, tt.wantFields...)
-			expected := observer.LoggedEntry{
-				Entry: zapcore.Entry{
-					Level:   tt.wantErrLevel,
-					Message: tt.wantInboundMsg,
-				},
-				Context: logContext,
-			}
-			assert.Equal(t, expected, getLog(), "Unexpected log entry written.")
-		})
-		t.Run(tt.desc+", stream outbound", func(t *testing.T) {
-			clientStream, err := mw.CallStream(context.Background(), sreq, newOutbound(tt))
-			checkErr(err)
-			logContext := append(
-				baseFields(),
-				zap.String("direction", string(_directionOutbound)),
-				zap.String("rpcType", "Streaming"),
-			)
-			logContext = append(logContext, tt.wantFields...)
-			if tt.err == nil {
-				assert.NotNil(t, clientStream, "Expected non-nil clientStream if call is successful.")
-			}
-			expected := observer.LoggedEntry{
-				Entry: zapcore.Entry{
-					Level:   tt.wantErrLevel,
-					Message: tt.wantOutboundMsg,
-				},
-				Context: logContext,
-			}
-			assert.Equal(t, expected, getLog(), "Unexpected log entry written.")
+			assert.Equal(t, expected, getLog(t), "Unexpected log entry written.")
 		})
 	}
+}
+
+func TestMiddlewareStreamingLogging(t *testing.T) {
+	defer stubTime()()
+	req := &transport.StreamRequest{
+		Meta: &transport.RequestMeta{
+			Caller:          "caller",
+			Service:         "service",
+			Transport:       "transport",
+			Encoding:        "raw",
+			Procedure:       "procedure",
+			Headers:         transport.NewHeaders().With("hello!", "goodbye!"),
+			ShardKey:        "shard-key",
+			RoutingKey:      "routing-key",
+			RoutingDelegate: "routing-delegate",
+		},
+	}
+
+	// helper function to creating logging fields for assertion
+	newZapFields := func(extraFields ...zapcore.Field) []zapcore.Field {
+		fields := []zapcore.Field{
+			zap.String("source", req.Meta.Caller),
+			zap.String("dest", req.Meta.Service),
+			zap.String("transport", req.Meta.Transport),
+			zap.String("procedure", req.Meta.Procedure),
+			zap.String("encoding", string(req.Meta.Encoding)),
+			zap.String("routingKey", req.Meta.RoutingKey),
+			zap.String("routingDelegate", req.Meta.RoutingDelegate),
+		}
+		return append(fields, extraFields...)
+	}
+
+	// create middleware
+	core, logs := observer.New(zapcore.DebugLevel)
+	infoLevel := zapcore.InfoLevel
+	mw := NewMiddleware(Config{
+		Logger:           zap.New(core),
+		Scope:            metrics.New().Scope(),
+		ContextExtractor: NewNopContextExtractor(),
+		Levels: LevelsConfig{
+			Default: DirectionalLevelsConfig{Success: &infoLevel},
+		},
+	})
+
+	// helper function to retrive observered logs, asserting the expected number
+	getLogs := func(t *testing.T, num int) []observer.LoggedEntry {
+		logs := logs.TakeAll()
+		require.Equal(t, num, len(logs), "expected exactly %d logs, got %v: %#v", num, len(logs), logs)
+
+		var entries []observer.LoggedEntry
+		for _, e := range logs {
+			// zero the time for easiy comparisons
+			e.Entry.Time = time.Time{}
+			entries = append(entries, e)
+		}
+		return entries
+	}
+
+	t.Run("success server", func(t *testing.T) {
+		stream, err := transport.NewServerStream(&fakeStream{request: req})
+		require.NoError(t, err)
+
+		err = mw.HandleStream(stream, &fakeHandler{
+			// send and receive messages in the handler
+			handleStream: func(stream *transport.ServerStream) {
+				err := stream.SendMessage(context.Background(), nil /*message*/)
+				require.NoError(t, err)
+				_, err = stream.ReceiveMessage(context.Background())
+				require.NoError(t, err)
+			}})
+		require.NoError(t, err)
+
+		logFields := func() []zapcore.Field {
+			return newZapFields(
+				zap.String("direction", string(_directionInbound)),
+				zap.String("rpcType", "Streaming"),
+				zap.Bool("successful", true),
+				zap.Skip(), // context extractor
+				zap.Skip(), // nil error
+			)
+		}
+
+		wantLogs := []observer.LoggedEntry{
+			{
+				// open stream
+				Entry: zapcore.Entry{
+					Message: _successStreamOpen,
+				},
+				Context: logFields(),
+			},
+			{
+				// send message
+				Entry: zapcore.Entry{
+					Message: _successfulStreamSend,
+				},
+				Context: logFields(),
+			},
+			{
+				// receive message
+				Entry: zapcore.Entry{
+					Message: _successfulStreamReceive,
+				},
+				Context: logFields(),
+			},
+			{
+				// close stream
+				Entry: zapcore.Entry{
+					Message: _successStreamClose,
+				},
+				Context: append(logFields(), zap.Duration("duration", 0)),
+			},
+		}
+
+		// log 1 - open stream
+		// log 2 - send message
+		// log 3 - receive message
+		// log 4 - close stream
+		gotLogs := getLogs(t, 4)
+		assert.Equal(t, wantLogs, gotLogs)
+	})
+
+	t.Run("error handler", func(t *testing.T) {
+		tests := []struct {
+			name string
+			err  error
+		}{
+			{
+				name: "client fault",
+				err:  yarpcerrors.InvalidArgumentErrorf("client err"),
+			},
+			{
+				name: "server fault",
+				err:  yarpcerrors.InternalErrorf("server err"),
+			},
+			{
+				name: "unknown fault",
+				err:  errors.New("unknown fault"),
+			},
+		}
+
+		for _, tt := range tests {
+			t.Run(tt.name, func(t *testing.T) {
+				stream, err := transport.NewServerStream(&fakeStream{request: req})
+				require.NoError(t, err)
+
+				err = mw.HandleStream(stream, &fakeHandler{err: tt.err})
+				require.Error(t, err)
+
+				fields := newZapFields(
+					zap.String("direction", string(_directionInbound)),
+					zap.String("rpcType", "Streaming"),
+					zap.Bool("successful", false),
+					zap.Skip(), // context extractor
+					zap.Error(tt.err),
+					zap.Duration("duration", 0),
+				)
+
+				wantLog := observer.LoggedEntry{
+					Entry: zapcore.Entry{
+						Message: _errorStreamClose,
+						Level:   zapcore.ErrorLevel,
+					},
+					Context: fields,
+				}
+
+				// The stream handler is only executed after a stream successfully connects
+				// with a client. Therefore the first streaming log will always be
+				// successful (tested in the previous subtest). We only care about the
+				// stream termination so we retrieve the last log.
+				//
+				// log 1 - open stream
+				// log 2 - close stream
+				gotLog := getLogs(t, 2)[1]
+				assert.Equal(t, wantLog, gotLog)
+			})
+		}
+	})
+
+	t.Run("error server - send and recv", func(t *testing.T) {
+		sendErr := errors.New("send err")
+		receiveErr := errors.New("receive err")
+
+		stream, err := transport.NewServerStream(&fakeStream{
+			request:    req,
+			sendErr:    sendErr,
+			receiveErr: receiveErr,
+		})
+		require.NoError(t, err)
+
+		err = mw.HandleStream(stream, &fakeHandler{
+			// send and receive messages in the handler
+			handleStream: func(stream *transport.ServerStream) {
+				err := stream.SendMessage(context.Background(), nil /*message*/)
+				require.Error(t, err)
+				_, err = stream.ReceiveMessage(context.Background())
+				require.Error(t, err)
+			}})
+		require.NoError(t, err)
+
+		fields := func() []zapcore.Field {
+			return newZapFields(
+				zap.String("direction", string(_directionInbound)),
+				zap.String("rpcType", "Streaming"),
+				zap.Bool("successful", false),
+				zap.Skip(), // context extractor
+			)
+		}
+
+		wantLogs := []observer.LoggedEntry{
+			{
+				// send message
+				Entry: zapcore.Entry{
+					Message: _errorStreamSend,
+					Level:   zapcore.ErrorLevel,
+				},
+				Context: append(fields(), zap.Error(sendErr)),
+			},
+			{
+				// receive message
+				Entry: zapcore.Entry{
+					Message: _errorStreamReceive,
+					Level:   zapcore.ErrorLevel,
+				},
+				Context: append(fields(), zap.Error(receiveErr)),
+			},
+		}
+
+		// We are only interested in the send and receive logs.
+		// log 1 - open stream
+		// log 2 - send message
+		// log 3 - receive message
+		// log 4 - close stream
+		gotLogs := getLogs(t, 4)[1:3]
+		assert.Equal(t, wantLogs, gotLogs)
+	})
+
+	t.Run("success client", func(t *testing.T) {
+		stream, err := mw.CallStream(context.Background(), req, fakeOutbound{})
+		require.NoError(t, err)
+		err = stream.SendMessage(context.Background(), nil /* message */)
+		require.NoError(t, err)
+		_, err = stream.ReceiveMessage(context.Background())
+		require.NoError(t, err)
+		require.NoError(t, stream.Close(context.Background()))
+
+		fields := func() []zapcore.Field {
+			return newZapFields(
+				zap.String("direction", string(_directionOutbound)),
+				zap.String("rpcType", "Streaming"),
+				zap.Bool("successful", true),
+				zap.Skip(), // context extractor
+				zap.Skip(), // nil error
+			)
+		}
+
+		wantLogs := []observer.LoggedEntry{
+			{
+				// stream open
+				Entry: zapcore.Entry{
+					Message: _successStreamOpen,
+				},
+				Context: fields(),
+			},
+			{
+				// stream send
+				Entry: zapcore.Entry{
+					Message: _successfulStreamSend,
+				},
+				Context: fields(),
+			},
+			{
+				// stream receive
+				Entry: zapcore.Entry{
+					Message: _successfulStreamReceive,
+				},
+				Context: fields(),
+			},
+			{
+				// stream close
+				Entry: zapcore.Entry{
+					Message: _successStreamClose,
+				},
+				Context: append(fields(), zap.Duration("duration", 0)),
+			},
+		}
+
+		// log 1 - open stream
+		// log 2 - send message
+		// log 3 - receive message
+		// log 4 - close stream
+		gotLogs := getLogs(t, 4)
+		assert.Equal(t, wantLogs, gotLogs)
+	})
+
+	t.Run("error client handshake", func(t *testing.T) {
+		clientErr := errors.New("client err")
+		_, err := mw.CallStream(context.Background(), req, fakeOutbound{err: clientErr})
+		require.Error(t, err)
+
+		fields := func() []zapcore.Field {
+			return newZapFields(
+				zap.String("direction", string(_directionOutbound)),
+				zap.String("rpcType", "Streaming"),
+				zap.Bool("successful", false),
+				zap.Skip(), // context extractor
+				zap.Error(clientErr),
+			)
+		}
+
+		wantLogs := []observer.LoggedEntry{
+			{
+				// stream open
+				Entry: zapcore.Entry{
+					Message: _errorStreamOpen,
+					Level:   zapcore.ErrorLevel,
+				},
+				Context: fields(),
+			},
+		}
+
+		// log 1 - open stream
+		gotLogs := getLogs(t, 1)
+		assert.Equal(t, wantLogs, gotLogs)
+	})
+
+	t.Run("error client - send recv close", func(t *testing.T) {
+		sendErr := errors.New("send err")
+		receiveErr := errors.New("receive err")
+		closeErr := errors.New("close err")
+
+		stream, err := mw.CallStream(context.Background(), req, fakeOutbound{
+			stream: fakeStream{
+				sendErr:    sendErr,
+				receiveErr: receiveErr,
+				closeErr:   closeErr,
+			}})
+		require.NoError(t, err)
+
+		err = stream.SendMessage(context.Background(), nil /* message */)
+		require.Error(t, err)
+		_, err = stream.ReceiveMessage(context.Background())
+		require.Error(t, err)
+		err = stream.Close(context.Background())
+		require.Error(t, err)
+
+		fields := func() []zapcore.Field {
+			return newZapFields(
+				zap.String("direction", string(_directionOutbound)),
+				zap.String("rpcType", "Streaming"),
+				zap.Bool("successful", false),
+				zap.Skip(), // context extractor
+			)
+		}
+
+		wantLogs := []observer.LoggedEntry{
+			{
+				// send message
+				Entry: zapcore.Entry{
+					Message: _errorStreamSend,
+					Level:   zapcore.ErrorLevel,
+				},
+				Context: append(fields(), zap.Error(sendErr)),
+			},
+			{
+				// receive message
+				Entry: zapcore.Entry{
+					Message: _errorStreamReceive,
+					Level:   zapcore.ErrorLevel,
+				},
+				Context: append(fields(), zap.Error(receiveErr)),
+			},
+			{
+				// close stream
+				Entry: zapcore.Entry{
+					Message: _errorStreamClose,
+					Level:   zapcore.ErrorLevel,
+				},
+				Context: append(fields(), zap.Error(closeErr), zap.Duration("duration", 0)),
+			},
+		}
+
+		// We are only interested in the send, receive and stream close logs
+		// log 1 - open stream
+		// log 2 - send message
+		// log 3 - receive message
+		// log 4 - close stream
+		gotLogs := getLogs(t, 4)[1:]
+		assert.Equal(t, wantLogs, gotLogs)
+	})
 }
 
 func TestMiddlewareMetrics(t *testing.T) {
@@ -491,8 +839,8 @@ func TestMiddlewareMetrics(t *testing.T) {
 	}
 
 	for _, tt := range tests {
-		validate := func(mw *Middleware, direction string) {
-			key, free := getKey(req, direction)
+		validate := func(mw *Middleware, direction string, rpcType transport.Type) {
+			key, free := getKey(req, direction, rpcType)
 			edge := mw.graph.getEdge(key)
 			free()
 			assert.Equal(t, int64(tt.wantCalls), edge.calls.Load())
@@ -516,7 +864,7 @@ func TestMiddlewareMetrics(t *testing.T) {
 				&transporttest.FakeResponseWriter{},
 				newHandler(tt),
 			)
-			validate(mw, string(_directionInbound))
+			validate(mw, string(_directionInbound), transport.Unary)
 		})
 		t.Run(tt.desc+", unary outbound", func(t *testing.T) {
 			mw := NewMiddleware(Config{
@@ -525,7 +873,7 @@ func TestMiddlewareMetrics(t *testing.T) {
 				ContextExtractor: NewNopContextExtractor(),
 			})
 			mw.Call(context.Background(), req, newOutbound(tt))
-			validate(mw, string(_directionOutbound))
+			validate(mw, string(_directionOutbound), transport.Unary)
 		})
 	}
 }
@@ -533,7 +881,7 @@ func TestMiddlewareMetrics(t *testing.T) {
 // getKey gets the "key" that we will use to get an edge in the graph.  We use
 // a separate function to recreate the logic because extracting it out in the
 // main code could have performance implications.
-func getKey(req *transport.Request, direction string) (key []byte, free func()) {
+func getKey(req *transport.Request, direction string, rpcType transport.Type) (key []byte, free func()) {
 	d := digester.New()
 	d.Add(req.Caller)
 	d.Add(req.Service)
@@ -543,6 +891,7 @@ func getKey(req *transport.Request, direction string) (key []byte, free func()) 
 	d.Add(req.RoutingKey)
 	d.Add(req.RoutingDelegate)
 	d.Add(direction)
+	d.Add(rpcType.String())
 	return d.Digest(), d.Free
 }
 
@@ -628,7 +977,7 @@ func TestMiddlewareSuccessSnapshot(t *testing.T) {
 			Body:            strings.NewReader("body"),
 		},
 		&transporttest.FakeResponseWriter{},
-		fakeHandler{nil, false},
+		fakeHandler{err: nil, applicationErr: false},
 	)
 	assert.NoError(t, err, "Unexpected transport error.")
 
@@ -641,6 +990,7 @@ func TestMiddlewareSuccessSnapshot(t *testing.T) {
 		"procedure":        "procedure",
 		"routing_delegate": "rd",
 		"routing_key":      "rk",
+		"rpc_type":         transport.Unary.String(),
 		"source":           "caller",
 	}
 	want := &metrics.RootSnapshot{
@@ -694,7 +1044,7 @@ func TestMiddlewareFailureSnapshot(t *testing.T) {
 			Body:            strings.NewReader("body"),
 		},
 		&transporttest.FakeResponseWriter{},
-		fakeHandler{fmt.Errorf("yuno"), false},
+		fakeHandler{err: fmt.Errorf("yuno"), applicationErr: false},
 	)
 	assert.Error(t, err, "Expected transport error.")
 
@@ -702,23 +1052,25 @@ func TestMiddlewareFailureSnapshot(t *testing.T) {
 	tags := metrics.Tags{
 		"dest":             "service",
 		"direction":        "inbound",
-		"transport":        "unknown",
 		"encoding":         "raw",
 		"procedure":        "procedure",
 		"routing_delegate": "rd",
 		"routing_key":      "rk",
+		"rpc_type":         transport.Unary.String(),
 		"source":           "caller",
+		"transport":        "unknown",
 	}
 	errorTags := metrics.Tags{
 		"dest":             "service",
 		"direction":        "inbound",
-		"transport":        "unknown",
 		"encoding":         "raw",
+		"error":            "unknown_internal_yarpc",
 		"procedure":        "procedure",
 		"routing_delegate": "rd",
 		"routing_key":      "rk",
+		"rpc_type":         transport.Unary.String(),
 		"source":           "caller",
-		"error":            "unknown_internal_yarpc",
+		"transport":        "unknown",
 	}
 	want := &metrics.RootSnapshot{
 		Counters: []metrics.Snapshot{
@@ -814,6 +1166,7 @@ func TestApplicationErrorSnapShot(t *testing.T) {
 				"procedure":        "procedure",
 				"routing_delegate": "rd",
 				"routing_key":      "rk",
+				"rpc_type":         transport.Unary.String(),
 				"source":           "caller",
 			}
 			errorTags := metrics.Tags{
@@ -824,6 +1177,7 @@ func TestApplicationErrorSnapShot(t *testing.T) {
 				"procedure":        "procedure",
 				"routing_delegate": "rd",
 				"routing_key":      "rk",
+				"rpc_type":         transport.Unary.String(),
 				"source":           "caller",
 				"error":            tt.errTag,
 			}